use two-way binding for add-link-modal
